+++ assembly/instructions/branch.ts
@@ -3,7 +3,7 @@ import { reg, u32SignExtend } from "./utils";
 
 // BRANCH_EQ_IMM
 export const branch_eq_imm: InstructionRun = (r, args, registers) => {
-  const b = u64(u32SignExtend(args.b));
+  const b = u32SignExtend(args.b);
   if (registers[reg(args.a)] === b) {
     return staticJump(r, args.c);
   }
@@ -12,7 +12,7 @@ export const branch_eq_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_NE_IMM
 export const branch_ne_imm: InstructionRun = (r, args, registers) => {
-  const b = u64(u32SignExtend(args.b));
+  const b = u32SignExtend(args.b);
   if (registers[reg(args.a)] !== b) {
     return staticJump(r, args.c);
   }
@@ -21,7 +21,7 @@ export const branch_ne_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_LT_U_IMM
 export const branch_lt_u_imm: InstructionRun = (r, args, registers) => {
-  const b = u64(u32SignExtend(args.b));
+  const b = u32SignExtend(args.b);
   if (registers[reg(args.a)] < b) {
     return staticJump(r, args.c);
   }
@@ -30,7 +30,7 @@ export const branch_lt_u_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_LE_U_IMM
 export const branch_le_u_imm: InstructionRun = (r, args, registers) => {
-  const b = u64(u32SignExtend(args.b));
+  const b = u32SignExtend(args.b);
   if (registers[reg(args.a)] <= b) {
     return staticJump(r, args.c);
   }
@@ -39,7 +39,7 @@ export const branch_le_u_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_GE_U_IMM
 export const branch_ge_u_imm: InstructionRun = (r, args, registers) => {
-  const b = u64(u32SignExtend(args.b));
+  const b = u32SignExtend(args.b);
   if (registers[reg(args.a)] >= b) {
     return staticJump(r, args.c);
   }
@@ -48,7 +48,7 @@ export const branch_ge_u_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_GT_U_IMM
 export const branch_gt_u_imm: InstructionRun = (r, args, registers) => {
-  const b = u64(u32SignExtend(args.b));
+  const b = u32SignExtend(args.b);
   if (registers[reg(args.a)] > b) {
     return staticJump(r, args.c);
   }
@@ -57,7 +57,7 @@ export const branch_gt_u_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_LT_S_IMM
 export const branch_lt_s_imm: InstructionRun = (r, args, registers) => {
-  if (i64(registers[reg(args.a)]) < u32SignExtend(args.b)) {
+  if (i64(registers[reg(args.a)]) < i64(u32SignExtend(args.b))) {
     return staticJump(r, args.c);
   }
   return ok(r);
@@ -65,7 +65,7 @@ export const branch_lt_s_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_LE_S_IMM
 export const branch_le_s_imm: InstructionRun = (r, args, registers) => {
-  if (i64(registers[reg(args.a)]) <= u32SignExtend(args.b)) {
+  if (i64(registers[reg(args.a)]) <= i64(u32SignExtend(args.b))) {
     return staticJump(r, args.c);
   }
   return ok(r);
@@ -73,7 +73,7 @@ export const branch_le_s_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_GE_S_IMM
 export const branch_ge_s_imm: InstructionRun = (r, args, registers) => {
-  if (i64(registers[reg(args.a)]) >= u32SignExtend(args.b)) {
+  if (i64(registers[reg(args.a)]) >= i64(u32SignExtend(args.b))) {
     return staticJump(r, args.c);
   }
   return ok(r);
@@ -81,7 +81,7 @@ export const branch_ge_s_imm: InstructionRun = (r, args, registers) => {
 
 // BRANCH_GT_S_IMM
 export const branch_gt_s_imm: InstructionRun = (r, args, registers) => {
-  if (i64(registers[reg(args.a)]) > u32SignExtend(args.b)) {
+  if (i64(registers[reg(args.a)]) > i64(u32SignExtend(args.b))) {
     return staticJump(r, args.c);
   }
   return ok(r);
